added login routes, authentication and authorization

+++ src/app/api/admin/deleteCourse/[id]/route.js
@@ -2,10 +2,15 @@
 import {connectToDb} from '@/app/lib/dbConnection/connect';
 import Course from '@/app/lib/models/Course';
 import { NextResponse } from 'next/server';
+import { isAuthenticated } from '@/app/lib/authentication/isAuthenticated';
+import Admin from '@/app/lib/models/Admin'
 
-export async function DELETE(req, {params}) {
+const handler = async(req, {params}) => {
     await connectToDb();
-  
+    const admin= Admin.findById(req.user.userId)
+    if(!admin){
+        return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    }
     try {
       const id = params.id
       const course = await Course.findByIdAndDelete(id);
@@ -20,3 +25,4 @@ export async function DELETE(req, {params}) {
       return NextResponse.json({ message: 'Error deleting course' }, { status: 500 });
     }
   }
+  export const DELETE = isAuthenticated(handler);
